move StackedSetMap into separate file
