+++ lib/Parser.js
@@ -6,11 +6,11 @@
 
 // Syntax: https://developer.mozilla.org/en/SpiderMonkey/Parser_API
 
-const util = require("util");
 const acorn = require("acorn-dynamic-import").default;
 const Tapable = require("tapable");
 const json5 = require("json5");
 const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
+const StackedSetMap = require("./util/StackedSetMap");
 
 function joinRanges(startRange, endRange) {
 	if(!endRange) return startRange;
@@ -37,77 +37,6 @@ const POSSIBLE_AST_OPTIONS = [{
 		dynamicImport: true
 	}
 }];
-
-class StackedSetMap {
-	constructor(defaultValue, parentStack) {
-		this.defaultValue = defaultValue;
-		this.stack = parentStack === undefined ? [] : parentStack.slice();
-		this.map = new Map();
-		this.stack.push(this.map);
-	}
-
-	add(item) {
-		this.map.set(item, true);
-	}
-
-	set(item, value) {
-		this.map.set(item, value);
-	}
-
-	delete(item) {
-		this.map.set(item, false);
-	}
-
-	has(item) {
-		return this.get(item, false);
-	}
-
-	get(item) {
-		const topValue = this.map.get(item);
-		if(typeof topValue !== "undefined")
-			return topValue;
-		for(var i = this.stack.length - 2; i >= 0; i--) {
-			const value = this.stack[i].get(item);
-			if(typeof value !== "undefined") {
-				this.map.set(item, value);
-				return value;
-			}
-		}
-		this.map.set(item, this.defaultValue);
-		return this.defaultValue;
-	}
-
-	_compress() {
-		this.map = new Map();
-		for(const data of this.stack) {
-			for(const pair of data) {
-				this.map.set(pair[0], pair[1]);
-			}
-		}
-		this.stack = [this.map];
-	}
-
-	asSet() {
-		this._compress();
-		return new Set(Array.from(this.map.entries()).filter(pair => pair[1]).map(pair => pair[0]));
-	}
-
-	createChild() {
-		return new StackedSetMap(this.defaultValue, this.stack);
-	}
-
-	get length() {
-		throw new Error("Parser.definitions is no longer an Array");
-	}
-
-	set length(value) {
-		throw new Error("Parser.definitions is no longer an Array");
-	}
-}
-
-StackedSetMap.prototype.push = util.deprecate(function(item) {
-	this.add(item);
-}, "Parser.definitions is no longer an Array: Use add instead.");
 
 class TrackingSet {
 	constructor(set) {
